u-boot-variscite: imx8mp-var-dart: Add experimental LVDS channel 1 patch
The experimental driver currently hardcodes the LVDS channel to channel 0.
This patch hardcodes it to channel 1.

Tested on imx8mp-var-som-symphony by changing lvds-channel0 to lvds-channel@1
and moving the display cable from J7 to J5.

Note that the LVDS patches are experimental. The patch from NXP that adds
the lvds driver makes changes that may not be compatible with other
machines and makes changes to the clocks that may not be compatible
with all displays. Use at your own risk.

Signed-off-by: Nate Drude <nate.d@variscite.com>